+++ rpackage/sentiment.ai/NEWS.md
@@ -1,3 +1,4 @@
 # sentiment.ai 0.1.0
 
 * Added a `NEWS.md` file to track changes to the package.
+* Tested for Microsoft R Open and CRAN R